app and testing finished

+++ src/domain/use-cases/save-file.use-case.test.ts
@@ -1,10 +1,30 @@
 import fs from 'fs';
 import { SaveFile } from './save-file.use-case';
 
+
+
 describe('SaveFileUseCase', () => {
 
-	afterEach(()=>{
-		fs.rmSync('outputs', { recursive: true });
+	const customOptions = {
+		fileContent: 'custom content',
+		fileDestination: 'custom-outputs/file-destination',
+		fileName: 'custom-table-name',
+	};
+
+	const customFilePath = `${customOptions.fileDestination}/${customOptions.fileName}.txt`;
+
+	// beforeEach(() => {
+	//   jest.clearAllMocks();
+	// });
+
+
+	afterEach(() => {
+		const outputFolderExists = fs.existsSync('outputs');
+		if (outputFolderExists) fs.rmSync('outputs', { recursive: true });
+
+		const customOutputFolderExists = fs.existsSync(customOptions.fileDestination);
+		if (customOutputFolderExists) fs.rmSync(customOptions.fileDestination, { recursive: true });
+
 	});
 
 	test('should save file with default values', () => {
@@ -16,12 +36,59 @@ describe('SaveFileUseCase', () => {
 		};
 
 		const result = saveFile.execute(options);
-		const checkFile = fs.existsSync('outputs/table.txt');
+		const fileExists = fs.existsSync(filePath); // ojo
 		const fileContent = fs.readFileSync(filePath, { encoding: 'utf-8' });
 
-		expect(result).toBeTruthy();
-		expect(checkFile).toBeTruthy();
+		expect(result).toBe(true);
+		expect(fileExists).toBe(true);
 		expect(fileContent).toBe(options.fileContent);
 
+
 	});
+
+
+	test('should save file with custom values', () => {
+
+		const saveFile = new SaveFile();
+
+		const result = saveFile.execute(customOptions);
+		const fileExists = fs.existsSync(customFilePath);
+		const fileContent = fs.readFileSync(customFilePath, { encoding: 'utf-8' });
+
+		expect(result).toBe(true);
+		expect(fileExists).toBe(true);
+		expect(fileContent).toBe(customOptions.fileContent);
+
+	});
+
+	test('should return false if directory could not be created', () => {
+
+		const saveFile = new SaveFile();
+		const mkdirSpy = jest.spyOn(fs, 'mkdirSync').mockImplementation(
+			() => { throw new Error('This is a custom error message from testing'); }
+		);
+
+		const result = saveFile.execute(customOptions);
+
+		expect(result).toBe(false);
+
+		mkdirSpy.mockRestore();
+
+	});
+
+
+	test('should return false if file could not be created', () => {
+
+		const saveFile = new SaveFile();
+		const writeFileSpy = jest.spyOn(fs, 'writeFileSync').mockImplementation(
+			() => { throw new Error('This is a custom writing error message'); }
+		);
+		const result = saveFile.execute({ fileContent: 'Hola' });
+
+		expect(result).toBe(false);
+		writeFileSpy.mockRestore();
+	});
+
+
+
 });
